[refac] large refactor on secret and env; now they're env/secret and env/options
