+++ backend/.env.sample
@@ -1,16 +1,21 @@
 # use this if you are using ./run-postgres-at-localhost:
 DSN="host=localhost user=user database=database password=password"
-# use this if you use docker-compose:
-# DSN="host=postgres user=user database=database password=password"
+# this line is not necessary if you are using docker-compose.yml
 
 # allow cors from 5173. set to empty string or delete to disable CORS.
 CORS_ORIGIN="http://localhost:5173"
 
-# if set to "true", backend serves ./static.
+# if set to "true", echo serves ./static
 ECHO_SERVES_FRONTEND_TOO=false
 
-# if set to "true", it will use token.json instead of querying to db.
-USE_TOKEN_JSON=true
+# if set to "file", it will use token.json instead of querying to db.
+# if set to "env", it will read from environment variables.
+# if set to "db", it will read matching token of user. use this in prod.
+TOKEN_SOURCE=db
+
+# if set to "env", it will construct credentials from env instead of reading credentials.json
+# if set to "none", credentials will be left nil.
+CREDENTIALS_SOURCE=file
 
 # if set to "true", it will use the same user for every request without authentication.
 STATIC_USER=true
