+++ backend/app/auth/auth_test.go
@@ -2,14 +2,13 @@ package auth_test
 
 import (
 	"context"
-	"os"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/ut-code/JourniCal/backend/app/auth"
 	cal "github.com/ut-code/JourniCal/backend/app/calendar"
-	"github.com/ut-code/JourniCal/backend/app/secret"
+	"github.com/ut-code/JourniCal/backend/app/env/secret"
 	"github.com/ut-code/JourniCal/backend/app/user"
 	"github.com/ut-code/JourniCal/backend/pkg/helper"
 	_ "github.com/ut-code/JourniCal/backend/pkg/tests/run-test-at-root"
@@ -35,7 +34,6 @@ var authURL string
 
 func init() {
 	var err error
-	os.Remove("./test.db")
 	db, err = gorm.Open(sqlite.Open("./test.db"))
 	helper.PanicOn(err)
 	err = db.AutoMigrate(&user.User{})
@@ -43,7 +41,7 @@ func init() {
 
 	config = secret.OAuth2Config
 	authURL = secret.AuthURL
-	token = secret.TokenFromJSON
+	token = secret.StaticToken
 }
 
 func TestBasicFunctionality(t *testing.T) {
